Feat: 메뉴바 UI 구현 (스타일 수정 필요)
